+++ src/components/voting/VotingCreditCounter.tsx
@@ -1,0 +1,31 @@
+import SmallTitle from "../SmallTitle";
+import Reset from "public/images/svg/Reset";
+type VotingCreditCounterProps = {
+  handleReset: any;
+  credits: any;
+};
+
+const VotingCreditCounter = ({
+  handleReset,
+  credits,
+}: VotingCreditCounterProps) => {
+  return (
+    <div className="rounded-lg bg-white border border-gray p-12 ml-6 text-center">
+      <SmallTitle text="VOICE CREDITS" />
+      <div className="text-4xl mt-2 min-w-[141.08px]">
+        <span className="font-semibold">{credits}</span>/100
+      </div>
+      <button
+        className={`transition-colors duration-200 ease-in-out transform  outline-none focus:outline-none flex flex-row items-center justify-center rounded-md font-semibold mx-auto border border-blue hover:bg-white focus:bg-white text-blue text-lg px-4 py-2 mt-5`}
+        onClick={() => handleReset()}
+      >
+        <span className="mr-3">
+          <Reset />
+        </span>
+        <span>Reset</span>
+      </button>
+    </div>
+  );
+};
+
+export default VotingCreditCounter;
