+++ src/components/voting/VotingHeaderButton.tsx
@@ -1,0 +1,24 @@
+import Add from "public/images/svg/Add";
+
+type VotingHeaderButtonProps = {
+  title: string;
+  link: string;
+};
+
+const VotingHeaderButton = ({ title, link }: VotingHeaderButtonProps) => {
+  return (
+    <a
+      className="transition-colors duration-200 ease-in-out transform outline-none focus:outline-none flex flex-row items-center justify-center rounded-md font-bold mx-auto border border-blue hover:bg-white focus:bg-white text-blue text-lg px-3 py-1 cursor-pointer"
+      href={link}
+      target="_blank"
+      rel="noreferrer"
+    >
+      <span className="mr-3">
+        <Add className="mb-1 fill-blue" />
+      </span>
+      <span>{title}</span>
+    </a>
+  );
+};
+
+export default VotingHeaderButton;
